requirements.txt
google-generativeai==0.3.2
streamlit==1.29.0
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.1.0